Factor out common LandlordNameAndContactTypes component.
